fix(chats): give the Chats tab the standard title and a soft scroll edge
The Chat tab drew "Chats" as a large flush headline inside its content
(Android parity — `screenTitleLarge`) with the new-chat button laid out
beside it, so the tab had to hide its navigation bar. A scroll edge effect
is painted by a bar's background, so with no bar there was nothing for the
list to fade against: rows met the status bar on the system's hard cut
line.

The title moves to `navigationTitle` with `.toolbarTitleDisplayMode(.inline)`
and the button to a `.topBarTrailing` toolbar item — the pattern the rest
of the app's screens already use. That puts the bar back, so the top edge
can take `scrollEdgeEffectStyle(.soft, for: .top)` and the rows blur out
under it. The button loses its `.liquidGlassButtonStyle` and the 32/44pt
availability split that went with it, since bar items get their glass from
the system on iOS 26.

`View.softScrollEdge(for:)` in FlipcashUI is that call behind the
availability check, so the app has one implementation of it.
`CurrencyInfoContentV2`'s private `SoftTopScrollEdge` was the first copy
and now routes through it.

`TipsIntroScreen` drops its `isEmbedded` parameter. `TipsScreen` sends an
embedded tab straight to the conversation list, so the intro is v1 only and
never drew a Chats title.
