+++ Flipcash/Core/Screens/Main/Tips/TipConversationsScreen.swift
@@ -15,57 +15,89 @@ struct TipConversationsScreen: View {
     @Environment(AppRouter.self) private var router
 
     /// See ``TipsScreen/isEmbedded`` — v2 drops the inline tip-card button and
-    /// leads with the large "Chats" title.
+    /// titles the screen "Chats".
     var isEmbedded: Bool = false
 
     var body: some View {
         let conversations = conversationController.conversations(of: .tipDm)
 
         Background(color: .backgroundMain) {
-            VStack(alignment: .leading, spacing: 0) {
-                if isEmbedded {
-                    ChatsTabTitle()
-                }
-
-                // v1 keeps its list even when empty — the "Show My Tip Card"
-                // button lives in it, so there is no blank state to fill.
-                if isEmbedded, conversations.isEmpty {
-                    NoChatsView()
-                } else {
-                    List {
-                        // v1 only — v2 reaches the tip card from its own tab.
-                        if !isEmbedded {
-                            Button("Show My Tip Card") {
-                                router.push(.tipcard)
-                            }
-                            .buttonStyle(.filled)
-                            .accessibilityIdentifier("show-my-tipcard-button")
-                            .listRowInsets(EdgeInsets(top: 16, leading: 20, bottom: 16, trailing: 20))
-                            .listRowBackground(Color.clear)
-                            .listRowSeparator(.hidden)
+            // v1 keeps its list even when empty — the "Show My Tip Card"
+            // button lives in it, so there is no blank state to fill.
+            if isEmbedded, conversations.isEmpty {
+                NoChatsView()
+            } else {
+                List {
+                    // v1 only — v2 reaches the tip card from its own tab.
+                    if !isEmbedded {
+                        Button("Show My Tip Card") {
+                            router.push(.tipcard)
                         }
+                        .buttonStyle(.filled)
+                        .accessibilityIdentifier("show-my-tipcard-button")
+                        .listRowInsets(EdgeInsets(top: 16, leading: 20, bottom: 16, trailing: 20))
+                        .listRowBackground(Color.clear)
+                        .listRowSeparator(.hidden)
+                    }
 
-                        ForEach(conversations) { conversation in
-                            TipConversationRow(conversation: conversation) {
-                                router.push(.tipConversation(conversation.id))
-                            }
+                    ForEach(conversations) { conversation in
+                        TipConversationRow(conversation: conversation) {
+                            router.push(.tipConversation(conversation.id))
                         }
                     }
-                    .listStyle(.plain)
-                    .scrollContentBackground(.hidden)
+                }
+                .listStyle(.plain)
+                .scrollContentBackground(.hidden)
+                // The list runs under the bar, so its top edge needs the soft
+                // fade rather than the system's default hard cut.
+                .softScrollEdge()
+            }
+        }
+        .navigationTitle(isEmbedded ? "Chats" : "Tips")
+        .toolbarTitleDisplayMode(.inline)
+        .toolbar {
+            if isEmbedded {
+                ToolbarItem(placement: .topBarTrailing) {
+                    NewChatButton()
                 }
             }
         }
-        .navigationTitle(isEmbedded ? "" : "Tips")
-        .toolbar(isEmbedded ? .hidden : .automatic, for: .navigationBar)
-        .toolbarTitleDisplayMode(.inline)
+    }
+}
+
+// MARK: - NewChatButton -
+
+/// The Chat tab's new-chat affordance, a bar item beside the "Chats" title.
+///
+/// The title used to be a large flush headline drawn in the content (Android
+/// parity — `screenTitleLarge`) with this button laid out next to it, which
+/// meant the tab hid its navigation bar. A scroll edge effect is drawn by the
+/// bar's background, so without a bar the list met the status bar on a hard
+/// line; the standard centred title puts the bar back and lets the list fade
+/// under it.
+private struct NewChatButton: View {
+
+    @Environment(AppRouter.self) private var router
+
+    var body: some View {
+        Button {
+            router.push(.usernameLookup)
+        } label: {
+            // Bar items get their glass from the system on iOS 26, so this
+            // carries no button style of its own.
+            Image.system(.plus)
+                .font(.appTextLarge)
+                .foregroundStyle(Color.textMain)
+        }
+        .accessibilityLabel("New chat")
+        .accessibilityIdentifier("new-chat-button")
     }
 }
 
 // MARK: - NoChatsView -
 
 /// The v2 Chats tab's empty state, shown until the first tip conversation
-/// exists — centred in the space between the "Chats" title and the tab bar, so
+/// exists — centred in the space between the navigation bar and the tab bar, so
 /// it sits at the same height as the tippable-profile intro on this tab.
 private struct NoChatsView: View {
